Added SQL Functionality
Made the main index page work with SQL as well as adding games and creating games. Had some troubles with getting users to interact properly with SQL but so far i can delete edit and create users with SQL but im having troubles with getting them to sign in with it.